Спринт 8: Добавить абстрактный класс TaskManagerTest и перенести общие тесты для классов имплементирующих TaskManager
